# docs/api/openapi.yml
openapi: 3.1.0

info:
  title: Ledger API
  description: Double-entry ledger & payment processing engine API
  version: 1.0.0

paths:
  /ledgers:
    post:
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerRequest"
      responses:
        '201':
          description: Ledger created successfuly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ledger"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  
  /ledgers/{ledgerId}:
    get:
      parameters:
        - name: ledgerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ledger found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ledger"
        '404':
          description: Ledger not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /ledgers/{ledgerId}/accounts:
    post:
      parameters:
        - name: ledgerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        '201':
          description: Account created successfuly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /ledgers/{ledgerId}/accounts/{accountId}:
    parameters:
      - name: ledgerId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: accountId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      responses:
        '200':
          description: Account found
          content:
            appplication/json:
              schema:
                $ref: "#/components/schemas/Account"
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

components:
  schemas:
    CreateLedgerRequest:
      type: object
      required:
        - name
        - metadata
      properties:
        name:
          type: string
        metadata:
          type: object
          description: Arbitrary metadata to include in ledger

    Ledger:
      type: object
      required:
        - id
        - name
        - metadata
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        metadata:
          type: object
          description: Arbitrary metadata to include in ledger

    CreateAccountRequest:
      type: object
      required:
        - name
        - currency
        - overdraftAllowed
        - metadata
      properties:
        name:
          type: string
        currency:
          $ref: "#/components/schemas/Currency"
        overdraftAllowed:
          type: boolean
        metadata:
          type: object
          description: Arbitrary metadata to include in account

    Account:
      type: object
      required:
        - id
        - ledgerId
        - name
        - currency
        - balance
        - availableBalance
        - overdraftAllowed
        - metadata
        - status
      properties:
        id:
          type: string
          format: uuid
        ledgerId:
          type: string
          format: uuid
        name:
          type: string
        currency:
          $ref: "#/components/schemas/Currency"
        balance:
          type: integer
          format: int64
        availableBalance:
          type: integer
          format: int64
        overdraftAllowed:
          type: boolean
        metadata:
          type: object
          description: Arbitrary metadata to include in account
        status:
          $ref: "#/components/schemas/AccountStatus"

    AccountStatus:
      type: string
      enum:
        - ACTIVE
        - FROZEN
        - BLOCKED

    Currency:
      type: string
      enum:
        - USD
        - EUR
        - BTC
        - ETH

    Error:
      type: object
      required:
        - type
        - message
        - details
      properties:
        type:
          $ref: "#/components/schemas/ErrorType"
        message:
          type: string
        details:
          type: array
          items:
            $ref: "#/components/schemas/ErrorDetail"

    ErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string

    ErrorType:
      type: string
      enum:
        - INTERNAL_ERROR
        - INVALID_REQUEST