# internal/account/query/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package query

import (
	"encoding/json"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Account struct {
	LedgerID         uuid.UUID
	ID               uuid.UUID
	Name             string
	Currency         string
	Balance          int64
	AvailableBalance int64
	OverdraftAllowed bool
	Metadata         json.RawMessage
	Status           string
	CreatedAt        pgtype.Timestamp
}