# internal/ledger/query/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package query

import (
	"encoding/json"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Ledger struct {
	ID        uuid.UUID
	Name      string
	Metadata  json.RawMessage
	CreatedAt pgtype.Timestamp
}